requirements.txt
# Core Framework
Flask==2.3.3
Flask-SqlAlchemy==3.0.5 # For database management
Flask-WTF==1.2.1
gunicorn==20.1.0        # WSGI HTTP server for running the Flask app
psycopg2-binary==2.9.7  # PostgreSQL database adapter
email-validator==2.1.1

# API
requests==2.31.0

# Email
Flask-Mail==0.9.1

# Administration
Flask-Admin==1.6.1

# User Management
Flask-Login==0.6.3      # For user session management
Werkzeug==2.3.7         # For password hashing and security

# Machine Learning Engine
torch==2.1.0            # Core PyTorch library
scikit-learn==1.3.1     # For machine learning algorithms

# Modules
python-dotenv==1.0.0    # For environment variable management
pandas==2.1.1           # For data manipulation and analysis (CSV handling)
qrcode==7.4.2           # For QR code generation
